add sidebar and create folder

+++ src/server/actions.ts
@@ -12,9 +12,8 @@ import { MUTATIONS } from "./db/queries";
 const utApi = new UTApi();
 export async function deleteFile(fileId: number) {
   const session = await auth();
-  if (!session.userId) {
-    throw new Error("Unauthorized");
-  }
+
+  if (!session.userId) return redirect("/sign-in");
 
   const [file] = await db
     .select()
@@ -25,6 +24,14 @@ export async function deleteFile(fileId: number) {
 
   if (!file) {
     throw new Error("File not found");
+  }
+
+  if (file.parent === null) {
+    throw new Error("Cannot delete root folder");
+  }
+
+  if (file.ownerId !== session.userId) {
+    throw new Error("Unauthorized");
   }
 
   const utapiResult = await utApi.deleteFiles([file.utKey]);
@@ -45,20 +52,14 @@ export async function deleteFile(fileId: number) {
 }
 
 export async function redirectToDrive() {
-  "use server";
-
   const session = await auth();
 
-  if (!session.userId) {
-    return redirect("/sign-in");
-  }
+  if (!session.userId) return redirect("/sign-in");
 
   return redirect("/drive");
 }
 
 export async function createNewDrive() {
-  "use server";
-
   const session = await auth();
 
   if (!session.userId) return redirect("/sign-in");
@@ -66,4 +67,24 @@ export async function createNewDrive() {
   const rootFolderId = await MUTATIONS.onboardUser(session.userId);
 
   return redirect(`/folder/${rootFolderId}`);
+}
+
+export async function createFolder(input: {
+  formData: FormData;
+  parent: number;
+  userId: string;
+}) {
+  await MUTATIONS.createFolder({
+    folder: {
+      name: input.formData.get("folderName") as string,
+      parent: input.parent,
+    },
+    userId: input.userId,
+  });
+
+  const c = await cookies();
+
+  c.set("force-refresh", JSON.stringify(Math.random()));
+
+  return { success: true };
 }
